// src/backend/interfaces.tsx
export interface Worklogs {
  Worklog: Worklog;
}

export interface RootTimeLogs {
  user: {
    User: User;
  };
  worklogs: Worklogs[];
}

export interface AvatarUrls {
  '48x48': string;
  '24x24': string;
  '16x16': string;
  '32x32': string;
}

export interface User {
  self: string;
  accountId: string;
  accountType: string;
  emailAddress?: string;
  avatarUrls: AvatarUrls;
  displayName: string;
  active: boolean;
  locale: string;
}
// ...
export interface Worklog {
  self: string;
  tempoWorklogId: number;
  issue: {
    self: string;
    id: number;
    issueName: string;
  };
  accountId: string;
  accountName: string;
  timeSpentSeconds: number;
  billableSeconds: number;
  issueStartDate: string;
  loggedTime: string;
  loggedDate: string;
  createdAt: string;
  updatedAt: string;
  author: {
    avatar: string;
    self: string;
    userID: string;
    displayName: string;
  };
  department: string;
  departmentLedgerCode: string;
}
export interface IssueBase {
  self: string;
  id: string;
}

export interface JiraResponseBase {
  self: string;
  metadata: Metadata;
}

export interface WorklogDictionary {
  userId: string;
  displayName: string;
  worklogs: FilteredWorklog[];
}
// ...
export interface IFilteredData {
  id: string;
  key: string;
  parent?: string | null;
}

export interface FilteredWorklog {
  issue: IFilteredData;
  worklog: Worklog;
  parent: string | null;
}
export interface FilteredWorklogData {
  logDate: string;
  worklog: FilteredWorklog;
}