// src/backend/mock-db.d.ts
export declare const UserDB: {
    displayName: string;
    accountId: string;
    accountType: string;
    emailAddress: string;
    avatarUrls: {
        "48x48": string;
        "24x24": string;
        "16x16": string;
        "32x32": string;
    };
    active: boolean;
    locale: string;
    self: string;
}[];
export declare const WorklogDB: import("./interfaces").Worklogs[][];